requirements.txt
py-cord
python-dotenv
qrcode
aiosqlite
requests